requirements.txt
pandas==2.2.1
matplotlib.pyplot==3.8.3
seaborn==0.13.2
streamlit==1.31.1
babel==2.14.0